requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
about-time==4.2.1
    # via alive-progress
alive-progress==3.1.5
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
blinker==1.6.2
    # via flask
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via flask
db-dtypes==1.1.1
    # via pandas-gbq
dirtyjson==1.0.8
    # via -r requirements.in
flask==2.3.2
    # via -r requirements.in
google-ads==21.2.0
    # via -r requirements.in
google-ai-generativelanguage==0.4.0
    # via google-generativeai
google-api-core[grpc]==2.11.0
    # via
    #   google-ads
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-generativeai
    #   pandas-gbq
google-api-python-client==2.88.0
    # via -r requirements.in
google-auth==2.19.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
    #   google-generativeai
    #   gspread
    #   pandas-gbq
    #   pydata-google-auth
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth-oauthlib==1.0.0
    # via
    #   google-ads
    #   gspread
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-aiplatform==1.25.0
    # via -r requirements.in
google-cloud-bigquery==3.11.4
    # via
    #   google-cloud-aiplatform
    #   pandas-gbq
google-cloud-bigquery-storage==2.22.0
    # via pandas-gbq
google-cloud-core==2.3.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.10.3
    # via google-cloud-aiplatform
google-cloud-storage==2.10.0
    # via google-cloud-aiplatform
google-crc32c==1.5.0
    # via google-resumable-media
google-generativeai==0.4.0
    # via -r requirements.in
google-resumable-media==2.5.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.60.0
    # via
    #   google-ads
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grapheme==0.6.0
    # via alive-progress
grpc-google-iam-v1==0.12.6
    # via google-cloud-resource-manager
grpcio==1.57.0
    # via
    #   google-ads
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.57.0
    # via
    #   google-ads
    #   google-api-core
gspread==5.10.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
numpy==1.24.3
    # via
    #   db-dtypes
    #   pandas
    #   pandas-gbq
    #   pyarrow
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via
    #   db-dtypes
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
pandas==2.0.3
    # via
    #   -r requirements.in
    #   db-dtypes
    #   pandas-gbq
pandas-gbq==0.19.2
    # via -r requirements.in
proto-plus==1.22.3
    # via
    #   google-ads
    #   google-ai-generativelanguage
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-resource-manager
protobuf==4.24.2
    # via
    #   google-ads
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-resource-manager
    #   google-generativeai
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==13.0.0
    # via
    #   db-dtypes
    #   pandas-gbq
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==2.6.3
    # via google-generativeai
pydantic-core==2.16.3
    # via pydantic
pydata-google-auth==1.8.2
    # via pandas-gbq
pyparsing==3.1.1
    # via httplib2
python-dateutil==2.8.2
    # via
    #   google-cloud-bigquery
    #   pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via google-ads
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
shapely==1.8.5.post1
    # via google-cloud-aiplatform
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
tqdm==4.66.2
    # via google-generativeai
typing-extensions==4.10.0
    # via
    #   google-generativeai
    #   pydantic
    #   pydantic-core
tzdata==2023.3
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
validators==0.22.0
    # via -r requirements.in
werkzeug==2.3.7
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
